+++ jest.config.js
@@ -12,7 +12,7 @@
 
 module.exports = {
   clearMocks: true,
-  coverageDirectory: "coverage",
-  coverageProvider: "v8",
-  testEnvironment: "node",
+  coverageDirectory: 'coverage',
+  coverageProvider: 'v8',
+  testEnvironment: 'node',
 };