+++ CHANGELOG.md
@@ -5,9 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!--
-## [Develop]
--->
+
+<!-- ## [Develop] -->
+
+## [2.0.0] (2020-11.28)
+### Added
+
+- added eslint and prettier as dev.-dep.
+- added tests for `lib/functions.js`
+- added `benchmark` as dev.-dep.
+- added simple benchmarks
+
+### Changed
+
+- **BREAKING**: optional parameter with object and in camelCase, `(arr, headers, start, increment_name, increment_step, custom_id_function)` to<br>`arr, parameter: {headers, start, incrementName, incrementStep, customIdFunction}` ([#4])
+- changed README.md to use optional parameter
+- removed `isObject` as dependency ([#6])
+- removed duplicate functions and add them to the file `lib/functions.js`
+- replaced `lodash` with single function-dep `lodash.isequal`
+
 
 ## [1.2.0] (2020-11-06)
 
@@ -48,7 +64,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - added tests for <b>only one</b> of the three branches from the formatter
 
 [#2]: https://github.com/LetsMelon/addIdToArray/issues/2
+[#4]: https://github.com/LetsMelon/addIdToArray/issues/4
+[#6]: https://github.com/LetsMelon/addIdToArray/issues/6
 
 [develop]: https://github.com/LetsMelon/addIdToArray/compare/main...develop
+[2.0.0]: https://github.com/LetsMelon/addIdToArray/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/LetsMelon/addIdToArray/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/LetsMelon/addIdToArray/compare/v1.0.0...v1.1.0